requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
defusedxml==0.7.1
Deprecated==1.2.18
ecdsa==0.19.1
Flask==3.1.2
Flask-Caching==2.3.1
Flask-Limiter==4.0.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==5.21.0
greenlet==3.2.4
httpie==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
multidict==6.7.0
mysql-connector-python==9.4.0
ordered-set==4.1.0
packaging==25.0
pyasn1==0.6.1
Pygments==2.19.2
PySocks==1.7.1
python-jose==3.5.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
